+++ pdata/ptrace/ptraceotlp/traces.go
@@ -115,10 +115,10 @@ func (tr Request) Traces() ptrace.Traces {
 	return ptrace.Traces(internal.NewTraces(tr.orig))
 }
 
-// Client is the client API for OTLP-GRPC Traces service.
+// GRPCClient is the client API for OTLP-GRPC Traces service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://godoc.org/google.golang.org/grpc#ClientConn.NewStream.
-type Client interface {
+type GRPCClient interface {
 	// Export ptrace.Traces to the server.
 	//
 	// For performance reasons, it is recommended to keep this RPC
@@ -126,12 +126,15 @@ type Client interface {
 	Export(ctx context.Context, request Request, opts ...grpc.CallOption) (Response, error)
 }
 
+// Deprecated: [0.61.0] Use GRPCClient instead
+type Client = GRPCClient
+
 type tracesClient struct {
 	rawClient otlpcollectortrace.TraceServiceClient
 }
 
 // NewClient returns a new Client connected using the given connection.
-func NewClient(cc *grpc.ClientConn) Client {
+func NewClient(cc *grpc.ClientConn) GRPCClient {
 	return &tracesClient{rawClient: otlpcollectortrace.NewTraceServiceClient(cc)}
 }
 
@@ -141,8 +144,8 @@ func (c *tracesClient) Export(ctx context.Context, request Request, opts ...grpc
 	return Response{orig: rsp}, err
 }
 
-// Server is the server API for OTLP gRPC TracesService service.
-type Server interface {
+// GRPCServer is the server API for OTLP gRPC TracesService service.
+type GRPCServer interface {
 	// Export is called every time a new request is received.
 	//
 	// For performance reasons, it is recommended to keep this RPC
@@ -150,13 +153,16 @@ type Server interface {
 	Export(context.Context, Request) (Response, error)
 }
 
-// RegisterServer registers the Server to the grpc.Server.
-func RegisterServer(s *grpc.Server, srv Server) {
+// Deprecated: [0.61.0] Use GRPCServer instead
+type Server = GRPCServer
+
+// RegisterServer registers the GRPCServer to the grpc.Server.
+func RegisterServer(s *grpc.Server, srv GRPCServer) {
 	otlpcollectortrace.RegisterTraceServiceServer(s, &rawTracesServer{srv: srv})
 }
 
 type rawTracesServer struct {
-	srv Server
+	srv GRPCServer
 }
 
 func (s rawTracesServer) Export(ctx context.Context, request *otlpcollectortrace.ExportTraceServiceRequest) (*otlpcollectortrace.ExportTraceServiceResponse, error) {